ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pymongo==4.6.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.0
